uh a lot

+++ chattriggers/say.py
@@ -1,12 +1,17 @@
 import chattrigger
-import os
+#import os
+
+import allowuselimited
 
 class Say(chattrigger.ChatTrigger):
 	
 	async def run(self, message, trigger, client):
 		#ownerid = int(os.environ.get("OWNER_ID"))
 		#if not message.author.id == ownerid: # only owner can use command
-		#	return
+			#return
+		if not allowuselimited.allowuselimited(message.author.id, client): # 2020-05-07 realized this would be a good idea for this
+			return
 		tosay = message.content[len(trigger):]
 		await message.channel.send(tosay)
 		await message.delete()
+
